asr_service/indic_conformer_service/requirements.txt
git+https://github.com/AshwinSankar17/NeMo-ai4b@nemo-v2
git+https://github.com/speechbrain/speechbrain.git@develop
torchaudio
pytorch-lightning==2.4.0
hydra-core==1.3.2
librosa==0.10.2.post1
sentencepiece==0.2.0
pandas==2.2.2
lhotse==1.27.0
editdistance==0.8.1
jiwer==3.0.4
pyannote.audio
webdataset==0.2.100
cython==0.29.37
pyyaml==6.0.2
argparse==1.4.0
onnxruntime==1.19.0
tqdm==4.66.5
transformers
huggingface_hub
tokenizers
datasets
inflect
soundfile
pydub
numpy<2.0
fastapi
uvicorn
python-multipart
loguru
IPython